move parse args to convert args test
