+++ tests/argparser/test_parser_utils.py
@@ -65,97 +65,3 @@ class TestParserUtils(CompareTestCase):
             InferFromFormatSubCommand.parsed_args_to_data_frame_handler_read_args(
                 {'input-format': 'excel', 'input': '-', 'encoding': 'utf-8', 'one_sheet': None}
             )
-
-    def test_parse_args_to_df_handler_convert_args(self):
-        with (self.input_folder / 'invoice_aito_schema.json').open() as f:
-            schema = json.load(f)
-
-        self.assertEqual(
-            {
-                'read_input': sys.stdin,
-                'write_output': sys.stdout,
-                'in_format': 'json',
-                'out_format': 'ndjson',
-                'read_options': {'encoding': 'utf-8'},
-                'convert_options': {},
-            },
-            ConvertFromFormatSubCommand.parsed_args_to_data_frame_handler_convert_args(
-                {'input-format': 'json', 'input': '-', 'encoding': 'utf-8', 'json': False, 'use_table_schema': None}
-            )
-        )
-
-        self.assertEqual(
-            {
-                'read_input': sys.stdin,
-                'write_output': sys.stdout,
-                'in_format': 'ndjson',
-                'out_format': 'json',
-                'read_options': {'encoding': 'utf-8'},
-                'convert_options': {},
-                'use_table_schema': schema
-            },
-            ConvertFromFormatSubCommand.parsed_args_to_data_frame_handler_convert_args(
-                {
-                    'input-format': 'ndjson',
-                    'input': '-', 'encoding': 'utf-8',
-                    'json': True,
-                    'use_table_schema': str(self.input_folder / 'invoice_aito_schema.json')
-                }
-            )
-        )
-
-        self.assertEqual(
-            {
-                'read_input': sys.stdin,
-                'write_output': sys.stdout,
-                'in_format': 'csv',
-                'out_format': 'ndjson',
-                'read_options': {'encoding': 'utf-8', 'delimiter': ',', 'decimal': '.'},
-                'convert_options': {},
-                'use_table_schema': schema
-            },
-            ConvertFromFormatSubCommand.parsed_args_to_data_frame_handler_convert_args(
-                {
-                    'input-format': 'csv',
-                    'input': '-',
-                    'encoding': 'utf-8',
-                    'json': False,
-                    'use_table_schema': str(self.input_folder / 'invoice_aito_schema.json'),
-                    'delimiter': ',',
-                    'decimal': '.'
-                }
-            )
-        )
-
-        self.assertEqual(
-            {
-                'read_input': self.input_folder / 'invoice.xlsx',
-                'write_output': sys.stdout,
-                'in_format': 'excel',
-                'out_format': 'ndjson',
-                'read_options': {'encoding': 'utf-8', 'sheet_name': 'sheet_name'},
-                'convert_options': {},
-            },
-            ConvertFromFormatSubCommand.parsed_args_to_data_frame_handler_convert_args(
-                {
-                    'input-format': 'excel',
-                    'input': str(self.input_folder / 'invoice.xlsx'),
-                    'encoding': 'utf-8',
-                    'json': False,
-                    'use_table_schema': None,
-                    'one_sheet': 'sheet_name'
-                }
-            )
-        )
-
-        with self.assertRaises(ParseError, msg = 'input must be a file path for excel files'):
-            ConvertFromFormatSubCommand.parsed_args_to_data_frame_handler_convert_args(
-                {
-                    'input-format': 'excel',
-                    'input': '-',
-                    'encoding': 'utf-8',
-                    'json': False,
-                    'use_table_schema': None,
-                    'sheet_name': None
-                }
-            )
